still having an issue with crashing, diagnosed it to line 112 for some reason trying to write to the last index in the array is what's causing a crash.

+++ include/mathhandler.h
@@ -1,3 +1,3 @@
-int calcSNMPPulses(unsigned long long Avg);
-CRGB calcPulseColor(unsigned long long Avg);
-CRGB calcPulseColor2(unsigned long long Avg);
+int calcSNMPPulses(unsigned long Avg);
+CRGB calcPulseColor(unsigned long Avg);
+CRGB calcPulseColor2(unsigned long Avg);
